Rename container to swift_container to match swift_object
